reference_backend/Gluco-guide-five_mods/Gluco-guide-five_mods/requirements.txt
fastapi
uvicorn
pandas
matplotlib
sqlalchemy
pydantic
googletrans==4.0.0-rc1
torch
scikit-learn
numpy
easyocr
python-multipart
sentence-transformers
pdfplumber
pytesseract
pdf2image
opencv-python
groq
python-dotenv
shap